+++ mvc/src/main/java/com/interface21/webmvc/servlet/HandlerAdapterRegistry.java
@@ -1,7 +1,7 @@
 package com.interface21.webmvc.servlet;
 
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerAdapter;
-import com.interface21.webmvc.servlet.mvc.tobe.annotation.AnnotationHandlerAdapter;
+import com.interface21.webmvc.servlet.mvc.HandlerAdapter;
+import com.interface21.webmvc.servlet.mvc.annotation.AnnotationHandlerAdapter;
 import java.util.ArrayList;
 import java.util.List;
